Move default config into separate file + some comments
