+++ rust-src/wallet_library/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod credential;
+pub mod default_wallet_config;
 pub mod identity;
 pub mod proofs;
 pub mod statement;
